Split CapitalTransferContractService into validator, email sender, and core
Pure move: CapitalTransferCreationValidator owns creation validation
with the concentration-limit and member-capital checks, the public
CapitalTransferEmailSender owns contract emails for the service and the
executor alike, and the contract service keeps the signing and state
lifecycle with ten dependencies instead of thirteen. A bean-name
collision with the execution-package validator forced the more precise
name, and a mislabeled pre-existing test is flagged in place.

God classes 12 -> 11, PMD violations 149 -> 147.
