+++ metrics/scorecard.json
@@ -7,30 +7,30 @@
   "deepNestingViolations": 0,
   "disabledTests": 5,
   "disconnectedClasses": 454,
-  "godClasses": 12,
+  "godClasses": 11,
   "lineCoverage": 82.86,
   "longClasses": 4,
   "meanInstability": 0.473,
   "moduleCycles": 0,
   "modulithViolations": 0,
   "mutationScore": 84.43,
-  "mutationScoreSliced": 85.19,
-  "mutationsSliced": 1283,
+  "mutationScoreSliced": 86.66,
+  "mutationsSliced": 1724,
   "mutationsTotal": 12781,
-  "pitestSlicedModules": 6,
-  "pmdViolations": 149,
+  "pitestSlicedModules": 8,
+  "pmdViolations": 147,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 35,
-    "CyclomaticComplexity": 77,
-    "GodClass": 12,
+    "CyclomaticComplexity": 76,
+    "GodClass": 11,
     "TooManyMethods": 5
   },
-  "slowest10TestClassesSeconds": 109.3,
+  "slowest10TestClassesSeconds": 110.2,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 162,
+  "suiteWallTimeSeconds": 171,
   "testClassesOver10Seconds": 5,
-  "top15ClassLines": 7092,
+  "top15ClassLines": 7050,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
