+++ src/main/java/ee/tuleva/onboarding/capital/transfer/CapitalTransferContractService.java
@@ -1,38 +1,26 @@
 package ee.tuleva.onboarding.capital.transfer;
 
-import static ee.tuleva.onboarding.capital.event.member.MemberCapitalEventType.*;
 import static ee.tuleva.onboarding.capital.transfer.CapitalTransferContractState.*;
 import static ee.tuleva.onboarding.event.TrackableEventType.CAPITAL_TRANSFER_STATE_CHANGE;
 import static ee.tuleva.onboarding.notification.OperationsNotificationService.Channel.CAPITAL_TRANSFER;
 import static ee.tuleva.onboarding.notification.email.EmailType.*;
 import static java.util.stream.Stream.concat;
 
-import com.microtripit.mandrillapp.lutung.view.MandrillMessage;
 import ee.tuleva.onboarding.auth.principal.AuthenticatedPerson;
-import ee.tuleva.onboarding.capital.CapitalRow;
-import ee.tuleva.onboarding.capital.CapitalService;
 import ee.tuleva.onboarding.capital.event.AggregatedCapitalEvent;
 import ee.tuleva.onboarding.capital.event.AggregatedCapitalEventRepository;
-import ee.tuleva.onboarding.capital.event.member.MemberCapitalEventType;
 import ee.tuleva.onboarding.capital.transfer.CapitalTransferContract.CapitalTransferAmount;
 import ee.tuleva.onboarding.capital.transfer.content.CapitalTransferContractContentService;
-import ee.tuleva.onboarding.epis.ContactDetailsService;
 import ee.tuleva.onboarding.event.TrackableEvent;
 import ee.tuleva.onboarding.notification.OperationsNotificationService;
-import ee.tuleva.onboarding.notification.email.EmailPersistenceService;
-import ee.tuleva.onboarding.notification.email.EmailService;
-import ee.tuleva.onboarding.notification.email.EmailType;
 import ee.tuleva.onboarding.signature.SignatureFile;
 import ee.tuleva.onboarding.user.User;
 import ee.tuleva.onboarding.user.UserService;
 import ee.tuleva.onboarding.user.member.Member;
 import ee.tuleva.onboarding.user.member.MemberService;
 import java.math.BigDecimal;
-import java.util.Base64;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -46,18 +34,15 @@ import org.springframework.transaction.annotation.Transactional;
 public class CapitalTransferContractService {
 
   private final CapitalTransferContractRepository contractRepository;
-  private final ActiveTransferCapital activeTransferCapital;
   private final UserService userService;
   private final MemberService memberService;
-  private final EmailService emailService;
-  private final EmailPersistenceService emailPersistenceService;
   private final CapitalTransferFileService capitalTransferFileService;
   private final CapitalTransferContractContentService contractContentService;
-  private final CapitalService capitalService;
   private final OperationsNotificationService notificationService;
-  private final ContactDetailsService contactDetailsService;
   private final ApplicationEventPublisher eventPublisher;
   private final AggregatedCapitalEventRepository aggregatedCapitalEventRepository;
+  private final CapitalTransferCreationValidator validator;
+  private final CapitalTransferEmailSender emailSender;
 
   public CapitalTransferContract create(
       AuthenticatedPerson sellerPerson, CreateCapitalTransferContractCommand command) {
@@ -65,7 +50,7 @@ public class CapitalTransferContractService {
     Member seller = sellerUser.getMemberOrThrow();
     Member buyer = memberService.getById(command.getBuyerMemberId());
 
-    validate(seller, buyer, command);
+    validator.validate(seller, buyer, command);
 
     BigDecimal currentOwnershipUnitPrice = getCurrentOwnershipUnitPrice();
     List<CapitalTransferAmount> transferAmountsWithUnitPrice =
@@ -92,131 +77,6 @@ public class CapitalTransferContractService {
     contract.setOriginalContent(contractContent);
 
     return contractRepository.save(contract);
-  }
-
-  private void validate(Member seller, Member buyer, CreateCapitalTransferContractCommand command) {
-    log.info(
-        "Validating command {} for seller {} and buyer {}", command, seller.getId(), buyer.getId());
-
-    if (isAmountsEmpty(command)) {
-      throw new IllegalArgumentException("No amounts specified");
-    }
-
-    if (!hasPositiveNonZeroAmountsPrices(command)) {
-      throw new IllegalArgumentException("Amounts or prices have negative or zero values");
-    }
-
-    if (!hasOnlyOneOfType(command)) {
-      throw new IllegalArgumentException("Duplicate types specified");
-    }
-
-    if (!hasOnlyLiquidatableTypes(command)) {
-      throw new IllegalArgumentException("Non-liquidatable capital types included in command");
-    }
-
-    if (seller.getId().equals(buyer.getId())) {
-      throw new IllegalArgumentException("Seller and buyer cannot be the same person.");
-    }
-
-    if (!hasEnoughMemberCapital(seller, command)) {
-      throw new IllegalStateException("Seller does not have enough member capital");
-    }
-
-    if (!isTransferWithinConcentrationLimit(buyer, command)) {
-      throw new IllegalStateException("Buyer would exceed concentration limit after transfer");
-    }
-  }
-
-  private boolean isTransferWithinConcentrationLimit(
-      Member buyer, CreateCapitalTransferContractCommand command) {
-    var user = buyer.getUser();
-
-    var totalMemberCapital =
-        capitalService.getCapitalRows(user.getMemberId()).stream()
-            .filter(event -> event.type() != UNVESTED_WORK_COMPENSATION)
-            .map(CapitalRow::getValue)
-            .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-    var memberCapitalBookValueBeingAcquired =
-        getCapitalBeingAcquiredInOtherTransfers(buyer).values().stream()
-            .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-    var memberCapitalBookValueToBeAcquired =
-        command.getTransferAmounts().stream()
-            .map(CapitalTransferAmount::bookValue)
-            .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-    var buyerMemberCapitalAfterPurchases =
-        totalMemberCapital
-            .add(memberCapitalBookValueBeingAcquired)
-            .add(memberCapitalBookValueToBeAcquired);
-
-    var concentrationLimit = capitalService.getCapitalConcentrationUnitLimit();
-    return concentrationLimit.compareTo(buyerMemberCapitalAfterPurchases) > 0;
-  }
-
-  public Map<MemberCapitalEventType, BigDecimal> getCapitalBeingAcquiredInOtherTransfers(
-      Member buyer) {
-    return activeTransferCapital.beingBoughtBy(buyer);
-  }
-
-  public Map<MemberCapitalEventType, BigDecimal> getCapitalBeingSoldInOtherTransfers(
-      Member seller) {
-    return activeTransferCapital.beingSoldBy(seller);
-  }
-
-  private boolean hasEnoughMemberCapital(
-      Member seller, CreateCapitalTransferContractCommand command) {
-
-    var capitalSoldInOtherTransfers = getCapitalBeingSoldInOtherTransfers(seller);
-
-    return command.getTransferAmounts().stream()
-        .allMatch(
-            transferAmount -> {
-              var totalSellerMemberCapitalOfType =
-                  capitalService.getCapitalRows(seller.getId()).stream()
-                      .filter(event -> event.type() == transferAmount.type())
-                      .map(CapitalRow::getValue)
-                      .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-              var capitalOfTypeBeingSoldInOtherTransfers =
-                  capitalSoldInOtherTransfers.getOrDefault(transferAmount.type(), BigDecimal.ZERO);
-              var totalCapitalToBeSold =
-                  transferAmount.bookValue().add(capitalOfTypeBeingSoldInOtherTransfers);
-
-              return totalSellerMemberCapitalOfType.compareTo(totalCapitalToBeSold) >= 0;
-            });
-  }
-
-  private boolean isAmountsEmpty(CreateCapitalTransferContractCommand command) {
-    return command.getTransferAmounts().isEmpty();
-  }
-
-  private boolean hasOnlyOneOfType(CreateCapitalTransferContractCommand command) {
-    return command.getTransferAmounts().stream()
-            .map(CapitalTransferAmount::type)
-            .collect(Collectors.toSet())
-            .size()
-        == command.getTransferAmounts().size();
-  }
-
-  private boolean hasPositiveNonZeroAmountsPrices(CreateCapitalTransferContractCommand command) {
-    return command.getTransferAmounts().stream()
-        .allMatch(
-            amount ->
-                amount.bookValue().compareTo(BigDecimal.ZERO) > 0
-                    && amount.price().compareTo(BigDecimal.ZERO) > 0);
-  }
-
-  private boolean hasOnlyLiquidatableTypes(CreateCapitalTransferContractCommand command) {
-    var typesToLiquidate =
-        command.getTransferAmounts().stream()
-            .map(CapitalTransferAmount::type)
-            .collect(Collectors.toSet());
-    var liquidatableTypes =
-        Set.of(CAPITAL_PAYMENT, WORK_COMPENSATION, MEMBERSHIP_BONUS, CAPITAL_ACQUIRED);
-
-    return liquidatableTypes.containsAll(typesToLiquidate);
   }
 
   public CapitalTransferContract getContract(Long id, User user) {
@@ -250,7 +110,8 @@ public class CapitalTransferContractService {
     contractRepository.save(contract);
     log.info("Contract {} signed by seller {}", contractId, contract.getSeller().getId());
 
-    sendContractEmail(contract.getBuyer().getUser(), CAPITAL_TRANSFER_BUYER_TO_SIGN, contract);
+    emailSender.sendContractEmail(
+        contract.getBuyer().getUser(), CAPITAL_TRANSFER_BUYER_TO_SIGN, contract);
   }
 
   public void signByBuyer(Long contractId, byte[] container, User user) {
@@ -303,7 +164,7 @@ public class CapitalTransferContractService {
 
     broadcastStateChangeEvent(contract::confirmPaymentByBuyer, contract, user);
     log.info("Payment confirmed by buyer for contract {}", id);
-    sendContractEmail(
+    emailSender.sendContractEmail(
         contract.getSeller().getUser(), CAPITAL_TRANSFER_CONFIRMED_BY_BUYER, contract);
     return contractRepository.save(contract);
   }
@@ -317,7 +178,7 @@ public class CapitalTransferContractService {
 
     broadcastStateChangeEvent(contract::confirmPaymentBySeller, contract, user);
     log.info("Payment confirmed by seller for contract {}.", id);
-    sendContractEmail(
+    emailSender.sendContractEmail(
         contract.getBuyer().getUser(), CAPITAL_TRANSFER_CONFIRMED_BY_SELLER, contract);
 
     try {
@@ -350,56 +211,6 @@ public class CapitalTransferContractService {
             Map.of("id", contract.getId(), "oldState", oldState, "newState", newState)));
   }
 
-  public void sendContractEmail(
-      User recipient, EmailType emailType, CapitalTransferContract contract) {
-    if (recipient.getEmail() == null) {
-      log.error("User {} has no email, not sending email {}", recipient.getId(), emailType);
-      return;
-    }
-
-    Map<String, Object> mergeVars =
-        Map.of(
-            "fname", recipient.getFirstName(),
-            "lname", recipient.getLastName(),
-            "sellerFirstName", contract.getSellerFirstName(),
-            "sellerLastName", contract.getSellerLastName(),
-            "sellerFullName", contract.getSellerFullName(),
-            "buyerFirstName", contract.getBuyerFirstName(),
-            "buyerLastName", contract.getBuyerLastName(),
-            "buyerFullName", contract.getBuyerFullName(),
-            "contractId", contract.getId());
-
-    var templateName = emailType.getTemplateName(getLanguage(recipient));
-    var attachments =
-        Set.of(CAPITAL_TRANSFER_CONFIRMED_BY_BUYER, CAPITAL_TRANSFER_CONFIRMED_BY_SELLER)
-                .contains(emailType)
-            ? contractAttachments(contract)
-            : List.<MandrillMessage.MessageContent>of();
-
-    MandrillMessage message =
-        emailService.newMandrillMessage(
-            recipient.getEmail(),
-            templateName,
-            mergeVars,
-            List.of("capital-transfer"),
-            attachments);
-
-    emailService
-        .send(recipient, message, templateName)
-        .ifPresent(
-            response ->
-                emailPersistenceService.save(
-                    recipient, response.getId(), emailType, response.getStatus()));
-  }
-
-  private String getLanguage(User user) {
-    //    var contactDetails = contactDetailsService.getContactDetails(user);
-    //    return contactDetails.getLanguagePreference() == ENG ? "en" : "et";
-
-    // hotfix: there is no authentication context when sending the email after board approval
-    return "et";
-  }
-
   private BigDecimal getCurrentOwnershipUnitPrice() {
     AggregatedCapitalEvent latestEvent =
         aggregatedCapitalEventRepository.findTopByOrderByDateDesc();
@@ -409,14 +220,5 @@ public class CapitalTransferContractService {
     }
 
     return latestEvent.getOwnershipUnitPrice();
-  }
-
-  private static List<MandrillMessage.MessageContent> contractAttachments(
-      CapitalTransferContract contract) {
-    MandrillMessage.MessageContent attachment = new MandrillMessage.MessageContent();
-    attachment.setName("liikmekapitali_avaldus" + contract.getId() + ".bdoc");
-    attachment.setType("application/bdoc");
-    attachment.setContent(Base64.getEncoder().encodeToString(contract.getDigiDocContainer()));
-    return List.of(attachment);
   }
 }
