+++ src/test/groovy/ee/tuleva/onboarding/capital/transfer/execution/CapitalTransferExecutionFlowTest.java
@@ -16,6 +16,7 @@ import ee.tuleva.onboarding.capital.transfer.CapitalTransferContract;
 import ee.tuleva.onboarding.capital.transfer.CapitalTransferContract.CapitalTransferAmount;
 import ee.tuleva.onboarding.capital.transfer.CapitalTransferContractRepository;
 import ee.tuleva.onboarding.capital.transfer.CapitalTransferContractService;
+import ee.tuleva.onboarding.capital.transfer.CapitalTransferEmailSender;
 import ee.tuleva.onboarding.user.member.Member;
 import java.math.BigDecimal;
 import java.util.List;
@@ -37,6 +38,7 @@ class CapitalTransferExecutionFlowTest {
   @Mock private CapitalTransferValidator validator;
   @Mock private CapitalTransferEventLinkRepository linkRepository;
   @Mock private CapitalTransferContractService capitalTransferContractService;
+  @Mock private CapitalTransferEmailSender emailSender;
 
   @Mock private CapitalTransferContract contract;
   @Mock private Member sellerMember;
@@ -57,7 +59,8 @@ class CapitalTransferExecutionFlowTest {
             memberCapitalEventRepository,
             validator,
             capitalTransferContractService,
-            linkRepository);
+            linkRepository,
+            emailSender);
 
     executionJob = new CapitalTransferExecutionJob(contractRepository, executor);
   }
@@ -154,12 +157,12 @@ class CapitalTransferExecutionFlowTest {
     verify(contract).executed();
     verify(contractRepository).save(contract);
 
-    verify(capitalTransferContractService, times(1))
+    verify(emailSender, times(1))
         .sendContractEmail(
             eq(contract.getSeller().getUser()),
             eq(CAPITAL_TRANSFER_APPROVED_BY_BOARD),
             eq(contract));
-    verify(capitalTransferContractService, times(1))
+    verify(emailSender, times(1))
         .sendContractEmail(
             eq(contract.getBuyer().getUser()),
             eq(CAPITAL_TRANSFER_APPROVED_BY_BOARD),
